libs-base: fixed config file path
Using "./" resulted in the following warning, because GSPrivateSymbolPath() returns only the library name without a full path:

GNUSTEP_CONFIG_FILE value ('') is not an absolute path.  Please rebuild GNUstep-base specifying a valid path to the config file.

Since runtime config files is not something we need to support we silence this warning by using "/" as config file path.